Update interface of AtomicFileStream

+++ src/SyncTrayzor/Utils/AtomicFileStream.cs
@@ -17,58 +17,61 @@ namespace SyncTrayzor.Utils
         private readonly string path;
         private readonly string tempPath;
 
-        public static AtomicFileStream New(string path, FileMode mode)
+        public static AtomicFileStream Open(string path, FileMode mode)
         {
-            return New(path, TempFilePath(path), mode, (mode == FileMode.Append ? FileAccess.Write : FileAccess.ReadWrite), FileShare.Read, DefaultBufferSize, FileOptions.None);
+            return Open(path, TempFilePath(path), mode, (mode == FileMode.Append ? FileAccess.Write : FileAccess.ReadWrite), FileShare.Read, DefaultBufferSize, FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, string tempFilePath, FileMode mode)
+        public static AtomicFileStream Open(string path, string tempFilePath, FileMode mode)
         {
-            return New(path, tempFilePath, mode, (mode == FileMode.Append ? FileAccess.Write : FileAccess.ReadWrite), FileShare.Read, DefaultBufferSize, FileOptions.None);
+            return Open(path, tempFilePath, mode, (mode == FileMode.Append ? FileAccess.Write : FileAccess.ReadWrite), FileShare.Read, DefaultBufferSize, FileOptions.None);
         }
 
-        public static  AtomicFileStream Bew(string path, FileMode mode, FileAccess access)
+        public static  AtomicFileStream Open(string path, FileMode mode, FileAccess access)
         {
-            return New(path, TempFilePath(path), mode, access, FileShare.Read, DefaultBufferSize, FileOptions.None);
+            return Open(path, TempFilePath(path), mode, access, FileShare.Read, DefaultBufferSize, FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, string tempFilePath, FileMode mode, FileAccess access)
+        public static AtomicFileStream Open(string path, string tempFilePath, FileMode mode, FileAccess access)
         {
-            return New(path, tempFilePath, mode, access, FileShare.Read, DefaultBufferSize, FileOptions.None);
+            return Open(path, tempFilePath, mode, access, FileShare.Read, DefaultBufferSize, FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, FileMode mode, FileAccess access, FileShare share)
+        public static AtomicFileStream Open(string path, FileMode mode, FileAccess access, FileShare share)
         {
-            return New(path, TempFilePath(path), mode, access, share, DefaultBufferSize, FileOptions.None);
+            return Open(path, TempFilePath(path), mode, access, share, DefaultBufferSize, FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, string tempFilePath, FileMode mode, FileAccess access, FileShare share)
+        public static AtomicFileStream Open(string path, string tempFilePath, FileMode mode, FileAccess access, FileShare share)
         {
-            return New(path, tempFilePath, mode, access, share, DefaultBufferSize, FileOptions.None);
+            return Open(path, tempFilePath, mode, access, share, DefaultBufferSize, FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize)
+        public static AtomicFileStream Open(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize)
         {
-            return New(path, TempFilePath(path), mode, access, share, bufferSize, FileOptions.None);
+            return Open(path, TempFilePath(path), mode, access, share, bufferSize, FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, bool useAsync)
+        public static AtomicFileStream Open(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, bool useAsync)
         {
-            return New(path, TempFilePath(path), mode, access, share, bufferSize, useAsync ? FileOptions.Asynchronous : FileOptions.None);
+            return Open(path, TempFilePath(path), mode, access, share, bufferSize, useAsync ? FileOptions.Asynchronous : FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, string tempFilePath, FileMode mode, FileAccess access, FileShare share, int bufferSize, bool useAsync)
+        public static AtomicFileStream Open(string path, string tempFilePath, FileMode mode, FileAccess access, FileShare share, int bufferSize, bool useAsync)
         {
-            return New(path, tempFilePath, mode, access, share, bufferSize, useAsync ? FileOptions.Asynchronous : FileOptions.None);
+            return Open(path, tempFilePath, mode, access, share, bufferSize, useAsync ? FileOptions.Asynchronous : FileOptions.None);
         }
 
-        public static AtomicFileStream New(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, FileOptions options)
+        public static AtomicFileStream Open(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, FileOptions options)
         {
-            return New(path, TempFilePath(path), mode, access, share, bufferSize, options);
+            return Open(path, TempFilePath(path), mode, access, share, bufferSize, options);
         }
 
-        public static AtomicFileStream New(string path, string tempPath, FileMode mode, FileAccess access, FileShare share, int bufferSize, FileOptions options)
+        public static AtomicFileStream Open(string path, string tempPath, FileMode mode, FileAccess access, FileShare share, int bufferSize, FileOptions options)
         {
+            if (access == FileAccess.Read)
+                throw new ArgumentException("If you're just opening the file for reading, AtomicFileStream won't help you at all");
+
             if (File.Exists(tempPath))
                 File.Delete(tempPath);
 
@@ -78,9 +81,19 @@ namespace SyncTrayzor.Utils
             return new AtomicFileStream(path, tempPath, mode, access, share, bufferSize, options);
         }
 
+        public static FileStream OpenWrite(string path)
+        {
+            return AtomicFileStream.Open(path, FileMode.OpenOrCreate, FileAccess.Write, FileShare.None);
+        }
+
         private AtomicFileStream(string path, string tempPath, FileMode mode, FileAccess access, FileShare share, int bufferSize, FileOptions options)
             : base(tempPath, mode, access, share, bufferSize, options)
         {
+            if (path == null)
+                throw new ArgumentNullException("path");
+            if (tempPath == null)
+                throw new ArgumentNullException("tempPath");
+
             this.path = path;
             this.tempPath = tempPath;
         }
